auth routes working

+++ db/migrations/20201026233653-create-product.js
@@ -14,7 +14,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.FLOAT(6, 2),
+        type: Sequelize.FLOAT,
       },
       images: {
         allowNull: false,
